foreported -r 4734:4736
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4737 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/function/blockvectorfunction/blockvectorfunction.hh
@@ -471,8 +471,8 @@ private:
   DofStorageType * dofArray_;  
   
   //! index 
-  mutable int count_;
-  mutable int idx_;
+  mutable size_t count_;
+  mutable size_t idx_;
 
 }; // end DofIteratorBlockVectorDiscreteFunction 
 template <class DiscreteFunctionSpaceImp>
